# Inline the large-weight heap to avoid interface conversions