Before:
```
BenchmarkAdd_Norm_100-8       	33945012	        35.3 ns/op	       8 B/op	       0 allocs/op
BenchmarkAdd_Norm_10000-8     	29864768	        35.3 ns/op	       8 B/op	       0 allocs/op
BenchmarkAdd_Norm_1000000-8   	 8718654	       209 ns/op	      48 B/op	       0 allocs/op
BenchmarkAdd_Exp_100-8        	34625696	        31.3 ns/op	       8 B/op	       0 allocs/op
BenchmarkAdd_Exp_10000-8      	39429486	        30.0 ns/op	       8 B/op	       0 allocs/op
BenchmarkAdd_Exp_1000000-8    	 9097611	       199 ns/op	      50 B/op	       0 allocs/op
```

After:
```
BenchmarkAdd_Norm_100-8       	37324060	        32.3 ns/op	       8 B/op	       0 allocs/op
BenchmarkAdd_Norm_10000-8     	41255592	        32.7 ns/op	       8 B/op	       0 allocs/op
BenchmarkAdd_Norm_1000000-8   	12280093	        98.6 ns/op	      30 B/op	       0 allocs/op
BenchmarkAdd_Exp_100-8        	36671206	        28.6 ns/op	       8 B/op	       0 allocs/op
BenchmarkAdd_Exp_10000-8      	41433784	        29.3 ns/op	       8 B/op	       0 allocs/op
BenchmarkAdd_Exp_1000000-8    	12819754	       107 ns/op	      29 B/op	       0 allocs/op
```